requirements.txt
streamlit
pandas
requests
beautifulsoup4
python-whois
fuzzywuzzy
openpyxl
tldextract
geopy
